+++ release/gff/gff_gurage/HISTORY.md
@@ -1,4 +1,7 @@
-# ጉራጌ (Gurage) Change History
+# ጉራጊና (Gurage) Change History
+
+## 2020-10-25 0.7
+* Updates to sync with the Gurage Legacy keyboard.
 
 ## 2020-07-12 0.6.12 (0.7 Beta 4)
 * Including fonts from https://github.com/raeytype/gurage-font-collection